pyproject.toml
[tool.poetry]
name = "scraping-job-ms"
version = "0.1.0"
description = "Modular Google Maps scraper · Producer-Consumer architecture with Strategy pattern for lead detection"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "src" },
    { include = "config" },
]

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.40.0"
aiohttp = "^3.9.0"
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.0"
streamlit = "^1.30.0"
pandas = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
